Passing CalloutPostCreatedSubscription variables correctly
